refactor: split out sync and plugin
