+++ package.json
@@ -69,10 +69,12 @@
     "push-dir": "^0.4.1"
   },
   "dependencies": {
-    "firebase": "^5.3.1",
     "memoizee": "^0.4.12",
     "object-path": "^0.11.4",
-    "picolog": "^1.0.4",
+    "picolog": "^1.0.4"
+  },
+  "peerDependencies": {
+    "firebase": "^5.3.1",
     "rxjs": "^6.2.2",
     "vue": "^2.5.17",
     "vuepress": "^0.13.1"
